+++ uisdk/src/test/java/com/karhoo/uisdk/screen/booking/booking/payment/adyen/AdyenPaymentPresenterTest.kt
@@ -8,7 +8,7 @@ import com.karhoo.sdk.api.model.adyen.AdyenPublicKey
 import com.karhoo.sdk.api.network.response.Resource
 import com.karhoo.sdk.api.service.payments.PaymentsService
 import com.karhoo.sdk.call.Call
-import com.karhoo.uisdk.screen.booking.booking.payment.PaymentMVP
+import com.karhoo.uisdk.screen.booking.booking.payment.BookingPaymentMVP
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.doNothing
@@ -27,7 +27,7 @@ class AdyenPaymentPresenterTest {
     private var paymentsService: PaymentsService = mock()
     private var userStore: UserStore = mock()
     private var savedPaymentInfo: SavedPaymentInfo = mock()
-    private var paymentView: PaymentMVP.View = mock()
+    private var paymentView: BookingPaymentMVP.View = mock()
     private var price: QuotePrice = mock()
     private val publicKeyCall: Call<AdyenPublicKey> = mock()
     private val publicKeyCaptor = argumentCaptor<(Resource<AdyenPublicKey>) -> Unit>()
